Mediawiki 1.23
It's not possible to call Specialpage anymore because of this error. I found a solution here: https://www.mail-archive.com/wikibugs-l@lists.wikimedia.org/msg53602.html. In the file AccessControlPanel.body.php I replaced the lines

       function AccessControlPanel() {
               SpecialPage::SpecialPage( 'AccessControlPanel', 'protect' );
       }

by

       function AccessControlPanel() {
               parent::__construct( 'AccessControlPanel', 'protect' );
       }

as suggested and now it works again. There should be a version control to choose the old behaviour in old mediawikis and change to the new in Mediawiki 1.23